Refactor Lifts and Ranks entities to use BigDecimal for weight; update LiftsDTO and RanksRepository with new methods; implement RankController and RankService for rank management

+++ src/main/java/com/a5ur4/goldengains/dtos/LiftsDTO.java
@@ -1,9 +1,11 @@
 package com.a5ur4.goldengains.dtos;
 
+import java.math.BigDecimal;
+
 public record LiftsDTO(
     Long id,
     String name,
-    Double weight,
+    BigDecimal weight,
     Integer reps,
     Integer sets,
     Long user
